requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohttp @ file:///croot/aiohttp_1707342283163/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1695717823297/work
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools @ file:///croot/cachetools_1713977085444/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1710350347627/work
cycler==0.12.1
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.51.0
frozenlist @ file:///croot/frozenlist_1698702560391/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///opt/conda/conda-bld/google-auth-oauthlib_1660687784486/work
grpcio @ file:///croot/grpc-suite_1701982571719/work
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.1
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.4.0
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.4.0
kiwisolver==1.4.5
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1701096859099/work
natsort==8.4.0
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib @ file:///croot/oauthlib_1679489621486/work
opencv-python==4.9.0.80
packaging==24.0
pillow @ file:///croot/pillow_1714398848491/work
protobuf==3.20.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///croot/pyopenssl_1708380408460/work
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.16.2
scikit-learn==1.3.2
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///croot/sympy_1701397643339/work
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1682445826165/work/tensorboard-2.12.1-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
threadpoolctl==3.5.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.2
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
urllib3 @ file:///croot/urllib3_1707770551213/work
Werkzeug @ file:///croot/werkzeug_1706210078083/work
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///croot/zipp_1704206909481/work
